requirements.txt
algoliasearch-django>=2.0,<3.0
django
djangorestframework
djangorestframework-simplejwt
django-dotenv
pyyaml
requests
django-cors-headers